+++ helperRoutines/mb_client_reader_v2.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 
 """
 MODBUS READER
@@ -25,9 +26,9 @@ def main():
 
     argparser = argparse.ArgumentParser(
         description="Universal MODBUS Client Reader")
-    argparser.add_argument('--ip',
+    argparser.add_argument('--host',
                            required=True,
-                           help='MODBUS Server Device IP',
+                           help='MODBUS Server Device IP/Name',
                            type=str
                            )
     argparser.add_argument('--port',
@@ -45,7 +46,7 @@ def main():
     _start_time = timer()
     try:
         mb_client = MODBUSClient(
-            ip=argparser.parse_args().ip,
+            host=argparser.parse_args().host,
             port=argparser.parse_args().port,
             debug=argparser.parse_args().debug
         )